docs(Text): move to component folder
